move out FilterWidget
