+++ catalog/app/containers/Search/FilterWidget.tsx
@@ -1,0 +1,179 @@
+import * as React from 'react'
+
+import * as FiltersUI from 'components/Filters'
+
+import * as SearchUIModel from './model'
+
+interface FilterWidgetProps<
+  P extends SearchUIModel.KnownPredicate = SearchUIModel.KnownPredicate,
+> {
+  state: SearchUIModel.PredicateState<P>
+  extents?: SearchUIModel.ExtentsForPredicate<P>
+  onChange: (state: SearchUIModel.PredicateState<P>) => void
+}
+
+export function NumberFilterWidget({
+  state,
+  extents,
+  onChange,
+}: FilterWidgetProps<SearchUIModel.Predicates['Number']>) {
+  const handleChange = React.useCallback(
+    (value: { min: number | null; max: number | null }) => {
+      onChange({ ...state, gte: value.min, lte: value.max })
+    },
+    [onChange, state],
+  )
+
+  // XXX: revisit this logic
+  const extentsComputed = React.useMemo(
+    () => ({
+      min: extents?.min ?? state.gte ?? 0,
+      max: extents?.max ?? state.lte ?? 0,
+    }),
+    [extents?.min, extents?.max, state.gte, state.lte],
+  )
+
+  return (
+    <FiltersUI.NumbersRange
+      extents={extentsComputed}
+      onChange={handleChange}
+      // XXX: add units for known filters
+      // unit={unit}
+      value={{ min: state.gte, max: state.lte }}
+    />
+  )
+}
+
+export function DatetimeFilterWidget({
+  state,
+  extents,
+  onChange,
+}: FilterWidgetProps<SearchUIModel.Predicates['Datetime']>) {
+  const fixedExtents = React.useMemo(
+    () => ({
+      min: extents?.min ?? new Date(),
+      max: extents?.max ?? new Date(),
+    }),
+    [extents?.min, extents?.max],
+  )
+
+  const fixedValue = React.useMemo(
+    () => ({ min: state.gte, max: state.lte }),
+    [state.gte, state.lte],
+  )
+
+  const handleChange = React.useCallback(
+    (v: { min: Date | null; max: Date | null }) => {
+      onChange({ ...state, gte: v.min, lte: v.max })
+    },
+    [onChange, state],
+  )
+
+  return (
+    <FiltersUI.DatesRange
+      extents={fixedExtents}
+      onChange={handleChange}
+      value={fixedValue}
+    />
+  )
+}
+
+const EMPTY_TERMS: string[] = []
+
+export function KeywordEnumFilterWidget({
+  state,
+  extents,
+  onChange,
+}: FilterWidgetProps<SearchUIModel.Predicates['KeywordEnum']>) {
+  const handleChange = React.useCallback(
+    (value: string[]) => {
+      onChange({ ...state, terms: value })
+    },
+    [onChange, state],
+  )
+  const availableValues = extents?.values ?? EMPTY_TERMS
+
+  return (
+    <FiltersUI.List
+      extents={availableValues}
+      onChange={handleChange}
+      value={state.terms}
+      placeholder="Find"
+    />
+  )
+}
+
+export function KeywordWildcardFilterWidget({
+  state,
+  onChange,
+}: FilterWidgetProps<SearchUIModel.Predicates['KeywordWildcard']>) {
+  const handleChange = React.useCallback(
+    (wildcard: string) => {
+      onChange({ ...state, wildcard })
+    },
+    [onChange, state],
+  )
+  // TODO: link to docs:
+  // https://www.elastic.co/guide/en/elasticsearch/reference/6.7/query-dsl-wildcard-query.html
+  return (
+    <FiltersUI.TextField
+      onChange={handleChange}
+      placeholder="Match against (wildcards supported)"
+      value={state.wildcard}
+    />
+  )
+}
+
+export function TextFilterWidget({
+  state,
+  onChange,
+}: FilterWidgetProps<SearchUIModel.Predicates['Text']>) {
+  const handleChange = React.useCallback(
+    (queryString: string) => {
+      onChange({ ...state, queryString })
+    },
+    [onChange, state],
+  )
+  // TODO: link to docs:
+  // https://www.elastic.co/guide/en/elasticsearch/reference/6.7/query-dsl-simple-query-string-query.html
+  return (
+    <FiltersUI.TextField
+      onChange={handleChange}
+      placeholder="Search for"
+      value={state.queryString}
+    />
+  )
+}
+
+type BooleanFilterValue = SearchUIModel.Untag<
+  SearchUIModel.PredicateState<SearchUIModel.Predicates['Boolean']>
+>
+
+export function BooleanFilterWidget({
+  state,
+  onChange,
+}: FilterWidgetProps<SearchUIModel.Predicates['Boolean']>) {
+  const handleChange = React.useCallback(
+    (value: BooleanFilterValue) => {
+      onChange({ ...state, ...value })
+    },
+    [onChange, state],
+  )
+  return <FiltersUI.BooleanFilter onChange={handleChange} value={state} />
+}
+
+const WIDGETS = {
+  Datetime: DatetimeFilterWidget,
+  Number: NumberFilterWidget,
+  Text: TextFilterWidget,
+  KeywordEnum: KeywordEnumFilterWidget,
+  KeywordWildcard: KeywordWildcardFilterWidget,
+  Boolean: BooleanFilterWidget,
+}
+
+export function FilterWidget(props: FilterWidgetProps) {
+  const Widget = WIDGETS[props.state._tag]
+  return <Widget {...(props as $TSFixMe)} />
+}
+
+export { FilterWidget as default }
